backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
langchain
langchain-core
langchain-community
langchain-huggingface
langgraph
langchain-groq
langchain-google-genai
langchain-openai
langchain-ollama
langchain-anthropic
deepagents
tavily
faiss-cpu
sentence-transformers
numpy
requests
httpx
python-multipart
PyYAML
